moved roundSigFigs to numbers, tests for getBaseProbability and roundSigFigs work but not for float32 edge case
